Rewrite of the shared memory / locking code - using ShmemInitStruct/LWLocks instead of custom code and semaphores.
From now on, the module has to be loaded using shared_preload_libraries to actually use it (it's possible to create the extension, but it's not possible to use it when loaded by a backend). But
that's actually a good thing, as this forces all backends to update the histogram.